Refactor liquidity mining
