+++ src/views/Homepage/BalanceHistoryTabs/TransactionHistoryTable/TransactionHistoryItems.tsx
@@ -9,6 +9,7 @@ import {
   IndexToken,
   MetaverseIndex,
 } from 'constants/tokens'
+
 import { AlchemyApiTransaction } from '../alchemyApi'
 
 import { TransactionHistoryItem } from './'
