+++ src/views/LiquidityMining/components/StakingModal.tsx
