+++ src/views/LiquidityMining/components/WarningMessage.tsx
