+++ src/views/LiquidityMining/useLiquidityMiningBalances.ts
@@ -12,10 +12,9 @@ import {
 } from 'constants/contractAddresses'
 import { GmiIndex } from 'constants/tokens'
 import { useNetwork } from 'hooks/useNetwork'
+import { useWallet } from 'hooks/useWallet'
 import { ERC20_ABI } from 'utils/abi/ERC20'
 import StakeRewardsABI from 'utils/abi/StakingRewards.json'
-
-import { useWallet } from './useWallet'
 
 /* Returns balance of ERC20 token address*/
 async function balanceOfAddress(
